Single parts brought together
Small fixes in classes FeatureVector, Match, Matcher, Sampler and VolumetricData.
Overwrote method ToString() for classes FeatureVector, Match and Point3D.
Prepared void Main() to the final construction (transforming missing).

+++ DataView/Sampler.cs
@@ -18,9 +18,14 @@ namespace DataView
 
             for (int i = 0; i < count; i++)
             {
-                int x = r.Next(rSphere, measures[0] - rSphere);
-                int y = r.Next(rSphere, measures[1] - rSphere);
-                int z = r.Next(rSphere, measures[2] - rSphere);
+                //DEBUG
+                //double x = r.Next(rSphere, measures[0] - rSphere)*d.GetXSpacing();
+                //double y = r.Next(rSphere, measures[1] - rSphere)*d.GetYSpacing();
+                //double z = r.Next(rSphere, measures[2] - rSphere)*d.GetZSpacing();
+
+                double x = r.Next(rSphere, measures[0] - rSphere);
+                double y = r.Next(rSphere, measures[1] - rSphere);
+                double z = r.Next(rSphere, measures[2] - rSphere);
 
                 points[i] = new Point3D(x, y, z);
             }
